media: camera: Add missing conftest.h
The camera VI driver was recently updated to fix the build for Linux
v6.10, but the change that fixed the build did not add the 'conftest.h'
header file which is needed. Add this header file to fix building the
driver for Linux v6.10.

Bug 4593750

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Change-Id: I11fdab89a1cb69552fbafbf4b4358561b71cac13
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/3150516
Reviewed-by: Brad Griffis <bgriffis@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>